// src/types/schemaTypes.d.ts
/* eslint-disable */
/**
 * This file was automatically generated by json-schema-to-typescript.
 * DO NOT MODIFY IT BY HAND. Instead, modify the source JSONSchema file,
 * and run json-schema-to-typescript to regenerate this file.
 */

export interface Schema {
  standards: {
    margin: number;
    padding: number;
    width: number;
    height: number;
    [k: string]: unknown;
  }[];
  definitions: {
    [k: string]: {
      x: number;
      y: number;
      [k: string]: unknown;
    };
  }[];
  [k: string]: unknown;
}